+++ Changelog.md
@@ -1,3 +1,7 @@
+##### 10.1.2
+* All: Added a "Forget Item" button to the Scrap merchant tab.
+* Dragonflight: Fixed issue that caused merchant "undo arrow" to appear while on the Scrap merchant tab.
+* 
 ##### 10.1.1
 * Dragonflight: Updated for new merchant stock UI.
 * Dragonflight: Added support for the new addon logos API and multiple-TOC format. (no Compartiment support yet).